pyproject.toml
[tool.poetry]
name = "dialogdoc"
version = "0.1.0"
description = ""
authors = ["y-tsuritani"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.4.0"
python-dotenv = "^1.0.0"
streamlit = "^1.29.0"
langchain = "^0.0.350"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
